TASK: Extract query and driver for Elastic v1
